+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RemoveTeam event.
 - An example showing how to use event sinks.
 - Example to demonstrate how to crate user defined events and metrics.
+- RegenerateStatistics event.
+- RegenerateAbilities event.
 
 ## [0.1.0] - 2020-02-08
 ### Added