+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-02-15
 ### Added
-- RemoveTeam event.
+- `RemoveTeam` event.
 - An example showing how to use event sinks.
 - Example to demonstrate how to crate user defined events and metrics.
-- RegenerateStatistics event.
-- RegenerateAbilities event.
+- `RegenerateStatistics` event.
+- `RegenerateAbilities` event.
+- `EntityId` now implements `Copy`.
 
 ## [0.1.0] - 2020-02-08
 ### Added